# documentatie/openAPI.yaml
openapi: 3.0.0
info:
  title: Opdrachtenbox API
  description: API voor het beheren van opdrachten voor studenten van de opleiding CMD
  version: 0.0.3
servers:
  - url: http://localhost:8080/
    description: Local development server
...
paths:
...
  /verify:
    get:
      tags:
        - Authenticatie
      summary: checkt authorization
      description: "Toegankelijk voor: iedereen"
      responses:
        "200":
          description: gebruiker is ingelogd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDto"
              examples:
                loggedOut:
                  value:
                    type: none
                    userId: null
                    businessId: null
                student:
                  value:
                    type: student
                    userId: 312
                    businessId: null
                supervisor:
                  value:
                    type: supervisor
                    userId: 312
                    businessId: 12
        "5XX":
          $ref: "#/components/responses/ServerError"
...
  /logout:
    post:
      tags:
        - Authenticatie
      summary: uitloggen
      description: |
        Deze endpoint wordt afgehandeld door OAuth

        Alleen toegankelijk voor: ingelogde gebruiker
      security:
        - sessionCookieAuth: []
      responses:
        "204":
          description: uitloggen successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: cookie header verwijderen
        "5XX":
          description: server fout
...
  /skills:
    get:
      tags:
        - Skills
      summary: ophalen van alle skills
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: skills succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSkillDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Skills
      summary: nieuwe skill toevoegen
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      requestBody:
        content:
          application/text:
            schema:
              type: string
              maxLength: 50
              example: adobe premiere pro
      responses:
        "201":
          description: skill succesvol toegevoegd
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetSkillDto"
        "400":
          description: skill heeft een foute opbouw, of bestaat al
        "401":
          description: gebruiker is niet ingelogd
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "5XX":
          description: server error
...
  /students:
    put:
      tags:
        - Studenten
      summary: updaten van student
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UpdateStudentDto"
      description: "Alleen toegankelijk voor: de student zelf"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: student successful aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker is niet geauthoriseerd om de pagina van de student aan te passen
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /students/email:
    get:
      tags:
        - Studenten
      summary: ophalen van emails van studenten
      parameters:
        - name: selection
          in: query
          required: true
          description: de selectie van studenten
          schema:
            type: integer
            format: int32
        - name: taskId
          in: query
          required: true
          description: de id van de taak
          schema:
            type: integer
            format: int32
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: email@email.com
                  description: email adres van student
          description: de email adressen van studenten zijn opgehaald
        "5XX":
          $ref: "#/components/responses/ServerError"
  /students/skill:
    put:
      tags:
        - Studenten
      summary: skill omschrijving van een student aanpassen
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetSkillWithDescriptionDto"
      description: "Alleen toegankelijk voor: de student zelf"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: skill succesvol aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de student rol
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /students/skills:
    put:
      tags:
        - Studenten
      summary: skills van een student aanpassen
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntArray"
      description: "Alleen toegankelijk voor: de student zelf"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: skills succesvol aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de student rol
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /students/{studentId}:
    get:
      tags:
        - Studenten
      summary: studenten ophalen
      parameters:
        - name: studentId
          in: path
          required: true
          description: id van het student
          schema:
            type: integer
            format: int32
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: student successful opgehaald
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDto"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "404":
          description: id is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /tasks/{projectId}:
    get:
      tags:
        - Taken
      summary: ophalen van alle Taken van een project
      description: "Alleen toegankelijk voor: ingelogde gebruiker"
      parameters:
        - name: projectId
          in: path
          required: true
          description: de id van het project
          schema:
            type: integer
            format: int32
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: rollen/klussen succesvol opgehaald
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskWithSkills"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de ingelogde gebruiker rol
        "404":
          description: project is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
    post:
      tags:
        - Taken
      summary: nieuwe taak toevoegen aan een project
      parameters:
        - name: projectId
          in: path
          required: true
          description: de id van het project
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskDto"
      description: "Alleen toegankelijk voor: opdrachtgever"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: taak succesvol toegevoegd
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "404":
          description: project is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
  /tasks/{taskId}/skills:
    put:
      tags:
        - Taken
      summary: skills van een Taken aanpassen
      parameters:
        - name: taskId
          in: path
          required: true
          description: de id van de taak
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntArray"
      description: "Alleen toegankelijk voor: opdrachtgever van het project"
      security:
        - sessionCookieAuth: []
      responses:
        "200":
          description: skills succesvol aangepast
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/UnAuthenticated"
        "403":
          description: gebruiker heeft niet de opdrachtgever rol
        "404":
          description: project is niet gevonden
        "5XX":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    sessionCookieAuth:
      type: apiKey
      in: cookie
      name: sessionToken
  responses:
    BadRequest:
      description: request body heeft een foute opbouw
    UnAuthenticated:
      description: gebruiker is niet ingelogd
    ServerError:
      description: server fout
  schemas:
...
    TaskWithSkills:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
          example: 12
        title:
          type: string
          maxLength: 50
          example: Taak titel
        description:
          type: string
          maxLength: 400
          example: Taak beschrijving
        totalNeeded:
          type: integer
          format: int32
          example: 3
        totalAccepted:
          type: integer
          format: int32
          example: 1
        totalRegistered:
          type: integer
          format: int32
          example: 2
        skills:
          type: array
          items:
            $ref: "#/components/schemas/GetSkillDto"
    GetSkillWithDescriptionDto:
      type: object
      properties:
        skill:
          $ref: "#/components/schemas/GetSkillDto"
        description:
          type: string
          maxLength: 400
          example: ik ben goed in adobe premiere pro
      required:
        - skill
        - description
    TaskDto:
      type: object
      properties:
        taskId:
          type: integer
          format: int32
          example: 12
        projectId:
          type: integer
          format: int32
          example: 5
        title:
          type: string
          maxLength: 50
          example: Nieuwe taak titel
        description:
          type: string
          maxLength: 400
          example: Nieuwe taak beschrijving
        totalNeeded:
          type: integer
          format: int32
          example: 3
      required:
        - taskId
        - projectId
        - title
        - description
        - totalNeeded
    GetSkillDto:
      description: een skill object
      type: object
      properties:
        skillId:
          type: integer
          format: int32
          example: 20
        name:
          type: string
          maxLength: 50
          example: adobe premiere pro
        isPending:
          type: boolean
          example: false
      required:
        - skillId
        - name
        - isPending
...
    FileDto:
      type: object
      properties:
        path:
          type: string
          example: /image.png
          maxLength: 255
    UpdateStudentDto:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          example: mijn naam
        description:
          type: string
          maxLength: 400
          example: ik ben een student met een passie voor ...
        profilePicture:
          type: string
          format: binary
        cv:
          type: string
          format: binary
    StudentDto:
      type: object
      properties:
        userId:
          type: integer
          format: int32
          example: 312
        username:
          type: string
          maxLength: 50
          example: mijn naam
        description:
          type: string
          maxLength: 400
          example: ik ben een student met een passie voor ...
        profilePicture:
          $ref: "#/components/schemas/FileDto"
        email:
          type: string
          example: name@student.han.nl
        cv:
          $ref: "#/components/schemas/FileDto"
        skills:
          type: array
          items:
            type: object
            properties:
              skill:
                $ref: "#/components/schemas/GetSkillDto"
              description:
                type: string
                maxLength: 400
                example: ik ben goed in adobe premiere pro
      required:
        - username
        - description
        - profilePicture
        - cv
        - skills
...
    VerificationDto:
      type: object
      properties:
        type:
          $ref: "#/components/schemas/VerificationType"
        userId:
          type: integer
          format: int32
          example: 312
        businessId:
          type: integer
          format: int32
          example: null
      required:
        - type

    # V
    VerificationType:
      description: authorisatie type
      type: string
      enum: [none, student, supervisor, teacher, invalid]
      example: student
...
    IntArray:
      type: array
      items:
        type: integer
        format: int32
      example: [1, 2, 3]